+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Fixes:
 
 - Blocking calls and assignments were generated in a wrong order in complex expressions
+- Don't output undefined arguments in call steps
 - Type annotation fixes in workflowslib
 
 ## Version 0.13.0 - 2025-10-12